src/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
aiogram==3.8.0
aiohttp==3.9.5
asyncpg==0.29.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
Telethon==1.36.0
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
requests==2.32.3
redis==5.0.7
python-dateutil==2.9.0.post0
pydantic_settings==2.8.1
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.3.0